Move zellij actions to shared package to expose as macros
Make the zellij helper actions (Directions, Resizes, Modes) public by
moving them to pkg/actions/tools/zellij so they're available as macros
for user specs and other completers.
