+++ completers/common/zellij_completer/cmd/action_movePane.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/common/zellij_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/zellij"
 	"github.com/spf13/cobra"
 )
 
@@ -19,5 +19,5 @@ func init() {
 	action_movePaneCmd.Flags().StringP("pane-id", "p", "", "Target a specific pane by ID (eg. terminal_1, plugin_2, or 3)")
 	actionCmd.AddCommand(action_movePaneCmd)
 
-	carapace.Gen(action_movePaneCmd).PositionalAnyCompletion(action.ActionDirections())
+	carapace.Gen(action_movePaneCmd).PositionalAnyCompletion(zellij.ActionDirections())
 }
