+++ completers/common/zellij_completer/cmd/action_resize.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/carapace-sh/carapace"
-	"github.com/carapace-sh/carapace-bin/completers/common/zellij_completer/cmd/action"
+	"github.com/carapace-sh/carapace-bin/pkg/actions/tools/zellij"
 	"github.com/spf13/cobra"
 )
 
@@ -20,7 +20,7 @@ func init() {
 	actionCmd.AddCommand(action_resizeCmd)
 
 	carapace.Gen(action_resizeCmd).PositionalCompletion(
-		action.ActionResizes(),
-		action.ActionDirections(),
+		zellij.ActionResizes(),
+		zellij.ActionDirections(),
 	)
 }
